% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/substr2.R
\name{ansi_state}
\alias{ansi_state}
\title{Compute String ANSI State at a Given Position}
\usage{
ansi_state(text, pos, type = "chars", lag, ends, tabs.as.spaces = FALSE,
  tab.stops = 8L)
}
\description{
Compute String ANSI State at a Given Position
}
