% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{dflt_term_cap}
\alias{dflt_term_cap}
\alias{dflt_css}
\title{Default Arg Helper Funs}
\usage{
dflt_term_cap()

dflt_css()
}
\value{
character to use as default value for \code{fansi} parameter.
}
\description{
Terminal capabilities are assumed to include bright and 256 color SGR codes.
24 bit color support is detected based on the \code{COLORTERM} environment
variable.
}
\details{
Default CSS may exceed or fail to cover the interline distance when two lines
have background colors.  To ensure lines are exactly touching use
inline-block, although that has its own issues.  Otherwise specify your own
CSS.
}
\seealso{
\code{\link{term_cap_test}}.
}
