% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tohtml.R
\name{esc_to_html}
\alias{esc_to_html}
\title{Convert ANSI CSI SGR Escape Sequence to HTML Equivalents}
\usage{
esc_to_html(x)
}
\arguments{
\item{x}{character vector}
}
\value{
a character vector with all escape sequences removed and any basic
ANSI CSI SGR escape sequences applied via SPAN html objects with specified
css styles.
}
\description{
Only the colors, background-colors, and basic styles (CSI SGR codes 1-9) are
translated.  Others are ignored.
}
\examples{
esc_to_html("hello\\033[31;42;1mworld\\033[m")
}
