% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand.R
\name{expand_sgr}
\alias{expand_sgr}
\title{Expand CSI SGR Sequences}
\usage{
expand_sgr(x, warn = getOption("fansi.warn"))
}
\arguments{
\item{x}{character vector to expand the SGR control sequences of.}

\item{warn}{TRUE (default) or FALSE, whether to warn when potentially
problematic \emph{Control Sequences} are encountered.  These could cause the
assumptions \code{fansi} makes about how strings are rendered on your display
to be incorrect, for example by moving the cursor (see \link{fansi}).}
}
\value{
\code{x}, with all SGRs expanded.
}
\description{
Converts compound SGR sequences into their component sequences for better
compatibility with \code{\link[=1]{crayon}}.  The closing sequence "ESC[0m" (or
"ESC[m") is expanded to match whatever styles are active at the point in
the string in which it appears, which includes styles accumulate from
previous vector elements.
}
\details{

}
\examples{
expand_sgr("hello\033[42;33m world")
expand_sgr("hello\033[42;33m world\033[m")
expand_sgr("\033[4mhello\033[42;33m world\033[m")
}
\seealso{
\code{\link{fansi}} for details on how \emph{Control Sequences} are
interpreted, particularly if you are getting unexpected results.
}
