% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fansi-package.R
\docType{package}
\name{fansi-package}
\alias{fansi-package}
\title{Simple Programmable NSE}
\description{
Counterparts to R string manipulation functions that preserve ANSI
CSI escape sequences while treating them as zero width.
}
