% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{fansi_lines}
\alias{fansi_lines}
\title{Colorize Character Vectors}
\usage{
fansi_lines(txt, step = 1)
}
\arguments{
\item{txt}{character to color}

\item{integer(1L)}{how quickly to step through the color palette}
}
\value{
character vector with each element colored
}
\description{
Color each element in input with one of the 256 color ANSI CSI SGR codes.
}
\examples{
NEWS <- readLines(file.path(R.home('doc'), 'NEWS'))
writeLines(fansi_lines(NEWS[1:20]))
writeLines(fansi_lines(NEWS[1:20], 8))
}
