% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{fansi_unicode_version}
\alias{fansi_unicode_version}
\title{Report Unicode Database Version In-use}
\usage{
fansi_unicode_version()
}
\value{
a string of the Unicode version.
}
\description{
Corresponds to the database used to generate character display width look-up
tables by \code{fansi}.
}
\examples{
fansi_unicode_version()
}
