% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{fwl}
\alias{fwl}
\title{Display Strings to Terminal}
\usage{
fwl(..., end = "<END>\\033[0m")
}
\arguments{
\item{...}{character vectors to display.}

\item{end}{character what to output after the primary inputs.}
}
\value{
whatever writeLines returns
}
\description{
Shortcut for \code{\link{writeLines}} with an additional terminating "ESC[0m".
}
\keyword{internal}
