% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/has.R
\name{has_csi}
\alias{has_csi}
\title{Checks Whether Character Vector Contains ANSI CSI Sequences}
\usage{
has_csi(x)
}
\arguments{
\item{x}{character}
}
\value{
logical of same length as \code{x}; NA values in \code{x} result in NA values
in return
}
\description{
...CSI definition...
}
