% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tohtml.R
\name{in_html}
\alias{in_html}
\title{Frame HTML in a Web Page And Display}
\usage{
in_html(x, css = "", display = TRUE)
}
\arguments{
\item{x}{character vector of html encoded strings}

\item{css}{character vector of css styles}

\item{display}{TRUE or FALSE, whether to display the resulting page in a
browser window.  If TRUE, will sleep for one second before returning, and
will delete the temporary file used to store the HTML.}
}
\value{
NULL invisibly
}
\description{
Helper function that takes HTML content and CSS, assembles a web page around
it, and uses \code{\link[=browseURL]{browseURL()}} to display it.  For testing purposes only.  No
effort is made to produce DOCTYPE, etc.
}
\examples{
txt <- "\033[31;42mHello \033[7mWorld\033[m"
writeLines(txt)
\dontrun{
in_html(txt) # spawns a browser window
}
readLines(in_html(txt, display=FALSE))
css <- "SPAN {text-decoration: underline;}"
readLines(in_html(txt, css=css, display=FALSE))
}
\seealso{
\code{\link[=make_styles]{make_styles()}}.
}
