% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tohtml.R
\name{make_styles}
\alias{make_styles}
\title{Generate CSS Mapping Classes to Colors}
\usage{
make_styles(classes, rgb.mix = diag(3))
}
\arguments{
\item{classes}{character vector of either 16, 32, or 512 class names.  See
\code{\link[=sgr_to_html]{sgr_to_html()}} for details, though note that unlike \code{\link[=sgr_to_html]{sgr_to_html()}} this
function only accepts character vectors.}

\item{rgb.mix}{3 x 3 numeric matrix to remix color channels.  Each column
corresponds to the channel mix of each channel in the output.  Intended
primarily to easily generate style sheets with different colors for demo
purposes.  Values that end up out of range are truncated into range.}
}
\value{
a character vector that can be used a style sheet.
}
\description{
Given a set of class names, produce the CSS that relates them to the default
8 bit colors.  This is a a helper function to generate style sheets for use
examples.
}
\examples{
class.8 <- do.call(paste, c(expand.grid(c("fg", "bg"), 0:7), sep="-"))
writeLines(make_styles(class.8))
## Use `rgb.mix` to remap color channels
mix <- matrix(
  c(
    0, 1, 0,  # red output is green input
    0, 0, 1,  # green output is blue input
    1, 0, 0   # blue output is red input
  ),
  nrow=3, byrow=TRUE
)
writeLines(make_styles(class.8, rgb.mix=mix))
}
\seealso{
Other HTML functions: 
\code{\link{in_html}()},
\code{\link{sgr_to_html}()}
}
\concept{HTML functions}
