% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{sgr_256}
\alias{sgr_256}
\title{Show 8 Bit ANSI CSI SGR Colors}
\usage{
sgr_256()
}
\value{
character vector with SGR codes with background color set as
themselves.
}
\description{
Generates text with each 8 bit SGR code (e.g. the "###" in "38;5;###") with
the background colored by itself, and the foreground in a contrasting color
and interesting color (we sacrifice some contrast for interest as this is
intended for demo rather than reference purposes).
}
\examples{
writeLines(sgr_256())
}
\seealso{
\code{\link[=make_styles]{make_styles()}}.
}
