% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/substr2.R
\name{state_esc}
\alias{state_esc}
\title{Compute String ANSI State at a Given Position}
\usage{
state_esc(text, pos, type = "chars", lag, ends, tabs.as.spaces = FALSE,
  tab.stops = 8L)
}
\description{
Compute String ANSI State at a Given Position
}
