% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strip.R
\name{strip_ansi}
\alias{strip_ansi}
\title{Strip Ansi Escape Sequences}
\usage{
strip_ansi(x)
}
\arguments{
\item{x}{character vector}
}
\value{
character vector of same length as x with ANSI escape sequences
stripped
}
\description{
...explain exactly what gets striped...
}
