% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sgr.R
\name{strip_sgr}
\alias{strip_sgr}
\title{Strip Control Sequences}
\usage{
strip_sgr(x, warn = getOption("fansi.warn"))
}
\arguments{
\item{x}{a character vector or object that can be coerced to such.}

\item{warn}{TRUE (default) or FALSE, whether to warn when potentially
problematic \emph{Control Sequences} are encountered.  These could cause the
assumptions \code{fansi} makes about how strings are rendered on your display
to be incorrect, for example by moving the cursor (see \code{\link[=fansi]{?fansi}}).
If the problematic sequence is a tab, you can use the \code{tabs.as.spaces}
parameter on functions that have it, or the \code{tabs_as_spaces} function, to
turn the tabs to spaces and resolve the warning that way.}
}
\value{
character vector of same length as x with ANSI escape sequences
stripped
}
\description{
This function is deprecated in favor of the \link[=strip_ctl]{\verb{_ctl} flavor}.  It
strips for CSI SGR and OSC-anchored URL sequences.
}
\examples{
## convenience function, same as `strip_ctl(ctl=c('sgr', 'url'))`
strip_sgr(string)
}
\keyword{internal}
