% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strtrim.R
\name{strtrim_esc}
\alias{strtrim_esc}
\title{ANSI CSI SGR aware version of \code{strtrim}}
\usage{
strtrim_esc(x, width, warn = getOption("fansi.warn"),
  term.cap = getOption("fansi.term.cap"))
}
\arguments{
\item{x}{a character vector, or an object which can be coerced to a
    character vector by \code{\link{as.character}}.}

\item{width}{Positive integer values: recycled to the length of \code{x}.}

\item{warn}{TRUE (default) or FALSE, whether to warn whether potentially
problematic escape sequences are encountered.  These could cause the
assumptions \code{fansi} is makes about how strings are rendered on your display
to be incorrect, for example by moving the cursor (see \link{string-parsing}).}

\item{term.cap}{character a vector of the capabilities of the terminal, can
be any combination "bright" (SGR codes 90-97, 100-107), "256" (SGR codes
starting with "38;5" or "48;5"), and "truecolor" (SGR codes starting with
"38;2" or "48;2"). Defaults to \code{c("bright", "256")}.  Changing this
parameter changes how \code{fansi} interprets escape sequences, so you should
ensure that it matches your terminal capabilities.}
}
\description{
ANSI CSI SGR aware version of \code{strtrim}
}
\examples{
strtrim_esc("\\033[42mHello world\\033[m", 6)
}
