% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strwrap.R
\name{strwrap_esc}
\alias{strwrap_esc}
\alias{strwrap2_esc}
\title{ANSI Escape Sequence Aware Versions of strwrap}
\usage{
strwrap_esc(x, width = 0.9 * getOption("width"), indent = 0, exdent = 0,
  prefix = "", simplify = TRUE, initial = prefix)

strwrap2_esc(x, width = 0.9 * getOption("width"), indent = 0, exdent = 0,
  prefix = "", simplify = TRUE, initial = prefix, wrap.always = TRUE,
  pad.end = " ", strip.spaces = FALSE, tabs.as.spaces = FALSE,
  tab.stops = 8L)
}
\arguments{
\item{x}{a character vector, or an object which can be converted to a
    character vector by \code{\link{as.character}}.}

\item{width}{a positive integer giving the target column for wrapping
    lines in the output.}

\item{indent}{a non-negative integer giving the indentation of the
    first line in a paragraph.}

\item{exdent}{a non-negative integer specifying the indentation of
    subsequent lines in paragraphs.}

\item{prefix}{a character string to be used as prefix for
    each line except the first, for which \code{initial} is used.}

\item{simplify}{a logical.  If \code{TRUE}, the result is a single
    character vector of line text; otherwise, it is a list of the same
    length as \code{x} the elements of which are character vectors of
    line text obtained from the corresponding element of \code{x}.
    (Hence, the result in the former case is obtained by unlisting that
    of the latter.)}

\item{initial}{a character string to be used as prefix for
    each line except the first, for which \code{initial} is used.}

\item{wrap.always}{TRUE or FALSE (default), whether to hard wrap at requested
width if no word breaks are detected within a line.}

\item{pad.end}{character(1L), a single character to use as padding at the
end of each line until the line is \code{width} wide.  This must be a printable
ASCII character, and defaults to a space.  Set to an empty string ("") to
leave line ends unpadded.}

\item{strip.spaces}{TRUE or FALSE (default), if TRUE, extraneous white spaces
(spaces, newlines, tabs) are removed in the same way as \link[base:strwrap]{base::strwrap}
does.}

\item{tabs.as.spaces}{TRUE (default) or FALSE  whether to convert tabs to
spaces.  This setting only has an effect if \code{strip.spaces == FALSE}.}

\item{tab.stops}{integer(1:n) indicating position of tab stops to use when
converting tabs to spaces.  If there are more tabs in a line than defined
tab stops the last tab stop is re-used.  For the purposes of applying tab
stops, each input line is considered a line and the character count begins
from the beginning of the input line.  Thus, if an input line is wrapped
the tab stops will not reset at the wrap point.  Additionally,\code{indent},
\code{exdent}, \code{initial}, and \code{prefix} will be ignored when computing tab
positions.}
}
\description{
Wraps strings to a specified width accounting for zero display width ANSI
escape sequences and control characters.  \code{strwrap_esc} is intended to
emulate \code{strwrap} exactly except with respect to the escape sequences, while
\code{strwrap2_esc} adds features and changes the processing of whitespace.
}
\details{
\code{strwrap2_esc} can convert tabs to spaces, pad strings up to \code{width}, and
hard-break words if single words are wider than \code{width}.

Unlike \link[base:strwrap]{base::strwrap}, both these functions will re-encode any non-ASCII
strings to UTF8 if they are not encoded in UTF8.
}
\examples{
hello.1 <- "hello \\033[41mred\\033[49m world"
hello.2 <- "hello\\t\\033[41mred\\033[49m\\tworld"

strwrap_esc(hello.1, 12)
strwrap_esc(hello.2, 12)

## With `strip.spaces=TRUE` strwrap2_esc is like strwrap_esc
strwrap2_esc(hello.2, 12, strip.spaces=TRUE)

## But by default it does not strip spaces
strwrap2_esc(hello.2, 12)

## You can convert tabs to spaces
strwrap2_esc(hello.2, 12, tabs.as.spaces=TRUE)

## If your display has 8 wide tab stops the following two
## outputs should look the same
writeLines(strwrap2_esc(hello.2, 80, tabs.as.spaces=TRUE))
writeLines(hello.2)

## tab stops are NOT auto-detected, but you may provide
## your own
strwrap2_esc(hello.2, 12, tabs.as.spaces=TRUE, tab.stops=c(6, 12))
}
\seealso{
\link{string-parsing} for important details on how strings are
interpreted and how character width is computed.
}
