% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unhandled.R
\name{unhandled_esc}
\alias{unhandled_esc}
\title{Identify Unhandled Escape Sequences}
\usage{
unhandled_esc(x)
}
\arguments{
\item{x}{character vector}
}
\value{
data frame with as many rows as there are unhandled escape
sequences and columns containing useful information for debugging the
problem.  See details.
}
\description{
Will record position and types of unhandled sequences in a character vector.
}
\details{
This is intended as a debugging function and as such is not optimized for
speed.

All byte offsets are computed on UTF-8 converted strings.  As such, you may
want to translate your character vector to UTF-8 if it contains non-ASCII
characters and it is in a different encoding.  This function will warn you if
it encounters such strings.

The return value is a data frame with five columns:
\itemize{
\item index: integer the index in \code{x} with the unhandled sequence
\item esc: character the unhandled escape sequence
\item start: integer the start position of the sequence (in characters)
\item stop: integer the start position of the sequence (in characters)
\item error: the reason why the sequence was not handled:
\itemize{
\item special: contains uncommon characters in ":<=>"
\item unknown: a substring with a value that does not correspond to a known
SGR code
\item non-SGR: a non-SGR CSI sequence
\item malformed: a malformed CSI sequence
\item non-CSI: a non-CSI escape sequence, i.e. one where the ESC is
followed by something other than "["
}
}
}
\examples{
string <- c(
  "\\033[41mhello world\\033[m", "foo\\033[22>m", "\\033[999mbar",
  "baz \\033[31#3m", "a\\033[31k", "hello\\033m world"
)
unhandled_esc(string)
}
\seealso{
\link{string-parsing} for details of how \code{fansi} interprets escape
sequences, \link{iconv} for how to translate strings across encodings.
}
