% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{html_esc}
\alias{html_esc}
\title{Escape Characters With Special HTML Meaning}
\usage{
html_esc(x)
}
\arguments{
\item{x}{character vector}
}
\value{
character vector consisting of \code{x}, but with the "<", ">", and "&"
characters replaced by their HTML entity codes.
}
\description{
This allows displaying strings that contain them in HTML without disrupting
the HTML.  It is assumed that the string to be escaped does not contain
actual HTML as this function would destroy it.
}
\examples{
html_esc("day > night")
html_esc("<SPAN>hello world</SPAN>")
}
