% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nchar.R
\name{nchar_sgr}
\alias{nchar_sgr}
\alias{nzchar_sgr}
\title{Control Sequence Aware Version of nchar}
\usage{
nchar_sgr(
  x,
  type = "chars",
  allowNA = FALSE,
  keepNA = NA,
  warn = getOption("fansi.warn", TRUE)
)

nzchar_sgr(x, keepNA = NA, warn = getOption("fansi.warn", TRUE))
}
\arguments{
\item{x}{a character vector or object that can be coerced to such.}

\item{type}{character(1L) partial matching \code{c("chars", "width")}, although
\code{type="width"} only works correctly with R >= 3.2.2.  See
\code{\link[base:nchar]{?nchar}}. With "width", the results might be affected by
locale changes, Unicode database updates, and logic changes for processing
of complex graphemes.  Generally you should not rely on a specific output
e.g. by embedding it in unit tests.  For the most part \code{fansi} (currently)
uses the internals of \code{base::nchar(type='width')}, but there are exceptions
and this may change in the future.}

\item{allowNA}{logical: should \code{NA} be returned for invalid
    multibyte strings or \code{"bytes"}-encoded strings (rather than
    throwing an error)?}

\item{keepNA}{logical: should \code{NA} be returned when
  \code{x} is \code{\link[base]{NA}}?  If false, \code{nchar()} returns
  \code{2}, as that is the number of printing characters used when
  strings are written to output, and \code{nzchar()} is \code{TRUE}.  The
  default for \code{nchar()}, \code{NA}, means to use \code{keepNA = TRUE}
  unless \code{type} is \code{"width"}.}

\item{warn}{TRUE (default) or FALSE, whether to warn when potentially
problematic \emph{Control Sequences} are encountered.  These could cause the
assumptions \code{fansi} makes about how strings are rendered on your display
to be incorrect, for example by moving the cursor (see \code{\link[=fansi]{?fansi}}).
If the problematic sequence is a tab, you can use the \code{tabs.as.spaces}
parameter on functions that have it, or the \code{tabs_as_spaces} function, to
turn the tabs to spaces and resolve the warning that way.}
}
\description{
These functions are deprecated in favor of the \link[=nchar_ctl]{\verb{_ctl} flavors}.
}
\keyword{internal}
