% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tohtml.R
\name{sgr_to_html}
\alias{sgr_to_html}
\alias{make_styles}
\title{Convert ANSI CSI SGR Escape Sequence to HTML Equivalents}
\usage{
sgr_to_html(
  x,
  warn = getOption("fansi.warn"),
  term.cap = getOption("fansi.term.cap"),
  classes = FALSE
)

make_styles(classes, rgb.mix = diag(3))
}
\arguments{
\item{x}{a character vector or object that can be coerced to character.}

\item{warn}{TRUE (default) or FALSE, whether to warn when potentially
problematic \emph{Control Sequences} are encountered.  These could cause the
assumptions \code{fansi} makes about how strings are rendered on your display
to be incorrect, for example by moving the cursor (see \link{fansi}).}

\item{term.cap}{character a vector of the capabilities of the terminal, can
be any combination of "bright" (SGR codes 90-97, 100-107), "256" (SGR codes
starting with "38;5" or "48;5"), and "truecolor" (SGR codes starting with
"38;2" or "48;2"). Changing this parameter changes how \code{fansi}
interprets escape sequences, so you should ensure that it matches your
terminal capabilities. See \link{term_cap_test} for details.}

\item{classes}{FALSE (default), TRUE, or character vector of either 16,
32, or 512 class names.  Character strings may only contain ASCII
characters corresponding to letters, numbers, the hyphen, or the
underscore.  It is the user's responsibility to provide values that are
legal class names.  \code{make_styles} only supports character vectors.
\itemize{
\item FALSE: All colors rendered as inline CSS styles.
\item TRUE: Each of the 256 basic colors is mapped to a class in form
"fansi-color-###" (or "fansi-bgcol-###" for background colors)
where "###" is a zero padded three digit number in 0:255.  Basic colors
specified with SGR codes 30-37 (or 40-47) map to 000:007, and bright ones
specified with 90-97 (or 100-107) map to 008:015.  8 bit colors specified
with SGR codes 38;5;### or 48;5;### map directly based on the value of
"###".  Implicitly, this maps the 8 bit colors in 0:7 to the basic
colors,  and those in 8:15 to the bright ones even though these are not
exactly the same when using inline styles.  "truecolor"s specified with
38;2;#;#;# or 48;2;#;#;# do not map to classes and are rendered as inline
styles instead.
\item character(16): The eight basic colors are mapped to the string values in
the vector, all others are rendered as inline CSS styles.  Basic colors
are mapped irrespective of whether they are encoded as the basic colors
or as 8 bit colors.  Sixteen elements are needed because there must be
eight classes for foreground colors, and 8 classes for background colors.
Classes should be ordered in ascending order of color number, with
foreground and background classes alternating starting with foreground
(see examples).
\item character(32): Like character(16), except the basic and bright colors are
mapped.
\item character(256): Like character(16), except the basic, bright, and all 8
bit colors are mapped.
}}

\item{rgb.mix}{3 x 3 numeric matrix to remix color channels.  Each column
corresponds to the channel mix of each channel in the output.  Intended
primarily to easily generate style sheets with different colors for demo
purposes.  Values that end up out of range are truncated into range.}
}
\value{
for \code{sgr_to_html}, a character vector with all escape sequences
removed and any basic ANSI CSI SGR escape sequences applied via SPAN html
objects with inline css styles (see details), for \code{make_styles} a
character vector that can be used a style sheet.
}
\description{
Only the colors, background-colors, and basic styles (CSI SGR codes 1-9) are
translated.  Others are dropped silently.
}
\details{
\code{make_styles} generates a style sheet to match provided classes to the
default 8 bit color mapping.  This is intended primarily for demo and test
purposes.
}
\note{
Non-ASCII strings are converted to and returned in UTF-8 encoding.
}
\examples{
sgr_to_html("hello\033[31;42;1mworld\033[m")
sgr_to_html("hello\033[31;42;1mworld\033[m", classes=TRUE)

## Generate some class names for basic colors
classes <- expand.grid(
  "myclass",
  c("fg", "bg"),
  c("black", "red", "green", "yellow", "blue", "magenta", "cyan", "white")
)
classes  # order is important!
classes <- do.call(paste, c(classes, sep="-"))
html <- sgr_to_html(
  "\033[94mhello\033[m \033[31;42;1mworld\033[m",
  classes=classes
)
## Only basic colors are mapped to classes; others styled inline
html

## Create a whole web page with a style sheet for 256 colors and
## the colors shown in a table.  For simplicity we'll only plot
## colors 16-231
class.256 <- do.call(paste, c(expand.grid(c("fg", "bg"), 0:255), sep="-"))
fg <- 231:16
bg <- rev(fg)  # reverse fg/bg so we can read the numbers
table <- matrix(
  sprintf("\033[38;5;\%d;48;5;\%dm\%s\033[m", fg, bg, format(bg)), 36
)
part.a <- do.call(paste0, c(split(table[1:18,], row(table[1:18,]))))
part.b <- do.call(paste0, c(split(table[-(1:18),], row(table[-(1:18),]))))

## Show colors on terminals that support 256 colors
part.a[1]
writeLines(part.a)
writeLines(part.b)

## Convert to HTML using classes instead of inline styles:
html.body <- sgr_to_html(c(part.a, part.b), classes=class.256)
writeLines(html.body[1])  # No inline colors
html.body <- paste0(html.body, collapse="\n")

## Web page template
page <- "
<html><head><style>\%s</style>\n
<body style='background-color: #EEEEEE;'>
<h3>\%s</h3>
<pre>\%s</pre></body></html>"

## Generate different style sheets.  We use `make_styles`
## for convenience but users should provide their own.
default <- make_styles(class.256)
mix <- matrix(c(.6,.2,.2, .2,.6,.2, .2,.2,.6), 3)
desaturated <- make_styles(class.256, mix)

writeLines(default[1:4])
writeLines(desaturated[1:4])
default <- paste0(default, collapse="\n")
desaturated <- paste0(desaturated, collapse="\n")

## Embed in HTML page and diplay; only CSS changing
html.0 <- sprintf(page, "", "Unstyled", html.body)
html.1 <- sprintf(page, default, "Default", html.body)
html.2 <- sprintf(page, desaturated, "Desaturated", html.body)

\dontrun{
f <- tempfile();
writeLines(html.0, f); browseURL(f); Sys.sleep(1);
writeLines(html.1, f); browseURL(f); Sys.sleep(1);
writeLines(html.2, f); browseURL(f); Sys.sleep(1);
unlink(f)
}
}
\seealso{
\link{fansi} for details on how \emph{Control Sequences} are
interpreted, particularly if you are getting unexpected results,
\code{\link[=set_knit_hooks]{set_knit_hooks()}} for how to use ANSI CSI styled text with knitr and HTML
output.
}
