% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strip.R
\name{strip_ctl}
\alias{strip_ctl}
\alias{strip_sgr}
\title{Strip ANSI Control Sequences}
\usage{
strip_ctl(x, strip = "all", warn = getOption("fansi.warn"))

strip_sgr(x, warn = getOption("fansi.warn"))
}
\arguments{
\item{x}{a character vector.}

\item{strip}{character, any combination of the following values (see details):
\itemize{
\item "nl": strip newlines
\item "c0": strip all other "C0" control characters (i.e. x01-x1f), except for
newlines and the actual ESC character
\item "sgr": strip ANSI CSI SGR sequences
\item "csi": strip all non-SGR csi sequences
\item "esc": strip all other escape sequences
\item "all": all of the above
}}

\item{warn}{TRUE (default) or FALSE, whether to warn when potentially
problematic escape sequences are encountered.  These could cause the
assumptions \code{fansi} makes about how strings are rendered on your display
to be incorrect, for example by moving the cursor (see \link{fansi}).}
}
\value{
character vector of same length as x with ANSI escape sequences
stripped
}
\description{
Removes \emph{Control Sequences} from strings.  By default it will
strip all known \emph{Control Sequences}, including ANSI CSI
sequences, two character sequences starting with ESC, and all C0 control
characters, including newlines.  You can fine tune this behavior with the
\code{strip} parameter.
}
\details{
The \code{strip} value contains the names of \strong{non-overlapping} subsets of the
known \emph{Control Sequences} (e.g. "csi" does not contain "sgr", and "c0" does
not contain newlines).  The one exception is "all" which means strip every
known sequence.  If you combine "all" with any other option then everything
\strong{but} that option will be stripped.
}
\examples{
string <- "hello\\033k\\033[45p world\\n\\033[31mgoodbye\\a moon"
strip_ctl(string)
strip_ctl(string, c("nl", "c0", "sgr", "csi", "esc")) # equivalently
strip_ctl(string, "sgr")
strip_ctl(string, c("c0", "esc"))
## everything but C0 controls (recall "nl" is not part of "c0")
## as far as the `strip` argument is concerned
strip_ctl(string, c("all", "nl", "c0"))

## convenience function, same as `strip_ctl(strip='sgr')`
strip_sgr(string)
}
\seealso{
\link{fansi} for details on how \emph{Control Sequences} are
interpreted, particularly if you are getting unexpected results.
}
