% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strip.R
\name{strip_esc}
\alias{strip_esc}
\title{Strip Ansi Escape Sequences}
\usage{
strip_esc(x, warn = getOption("fansi.warn"))
}
\arguments{
\item{x}{character vector}
}
\value{
character vector of same length as x with ANSI escape sequences
stripped
}
\description{
...explain exactly what gets striped...
}
