% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/substr2.R
\name{substr_esc}
\alias{substr_esc}
\title{Alternate substr version}
\usage{
substr_esc(x, start, stop, type = "chars", round = "first",
  tabs.as.spaces = FALSE, tab.stops = 8L)
}
\description{
Alternate substr version
}
