% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{tabs_as_spaces}
\alias{tabs_as_spaces}
\title{Replace Tabs With Corresponding Spaces}
\usage{
tabs_as_spaces(x, tab.stops = getOption("fansi.tab.stops"))
}
\arguments{
\item{x}{character vector to replace tabs in.}

\item{tab.stops}{integer distance between tab stops, the last tab stop
repeats as necessary.}
}
\value{
character, \code{x} with tabs replaced by spaces, with elements
possibly converted to UTF-8.
}
\description{
Finds horizontal tab characters (0x09) in a string and replaces them with the
spaces that produce the same horizontal offset.
}
\details{
Since we do not know of a reliable cross platform means of detecting tab
stops you will need to provide them yourself if you are using anything
outside of the standard tab stop every 8 characters that is the default.
}
\examples{
string <- '1\\t12\\t123\\t1234\\t12345678'
tabs_as_spaces(string)
writeLines(
  c(
    '-------|-------|-------|-------|-------|',
    tabs_as_spaces(string)
) )
writeLines(
  c(
    '-|--|--|--|--|--|--|--|--|--|--|',
    tabs_as_spaces(string, tab.stops=c(2, 3))
) )
writeLines(
  c(
    '-|--|-------|-------|-------|',
    tabs_as_spaces(string, tab.stops=c(2, 3, 8))
) )
}
\seealso{
\link{string-parsing} for important details on how strings are
interpreted
}
