% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{tabs_as_spaces}
\alias{tabs_as_spaces}
\title{Replace Tabs With Corresponding Spaces}
\usage{
tabs_as_spaces(x, tab.stops = getOption("fansi.tab.stops"),
  warn = getOption("fansi.warn"), term.cap = getOption("fansi.term.cap"))
}
\arguments{
\item{x}{character vector to replace tabs in.}

\item{tab.stops}{integer(1:n) indicating position of tab stops to use when
converting tabs to spaces.  If there are more tabs in a line than defined
tab stops the last tab stop is re-used.  For the purposes of applying tab
stops, each input line is considered a line and the character count begins
from the beginning of the input line.}

\item{warn}{TRUE (default) or FALSE, whether to warn whether potentially
problematic escape sequences are encountered.  These could cause the
assumptions \code{fansi} is makes about how strings are rendered on your display
to be incorrect, for example by moving the cursor (see \link{string-parsing}).}

\item{term.cap}{character a vector of the capabilities of the terminal, can
be any combination "bright" (SGR codes 90-97, 100-107), "256" (SGR codes
starting with "38;5" or "48;5"), and "truecolor" (SGR codes starting with
"38;2" or "48;2"). Defaults to \code{c("bright", "256")}.  Changing this
parameter changes how \code{fansi} interprets escape sequences, so you should
ensure that it matches your terminal capabilities.}
}
\value{
character, \code{x} with tabs replaced by spaces, with elements
possibly converted to UTF-8.
}
\description{
Finds horizontal tab characters (0x09) in a string and replaces them with the
spaces that produce the same horizontal offset.
}
\details{
Since we do not know of a reliable cross platform means of detecting tab
stops you will need to provide them yourself if you are using anything
outside of the standard tab stop every 8 characters that is the default.
}
\examples{
string <- '1\\t12\\t123\\t1234\\t12345678'
tabs_as_spaces(string)
writeLines(
  c(
    '-------|-------|-------|-------|-------|',
    tabs_as_spaces(string)
) )
writeLines(
  c(
    '-|--|--|--|--|--|--|--|--|--|--|',
    tabs_as_spaces(string, tab.stops=c(2, 3))
) )
writeLines(
  c(
    '-|--|-------|-------|-------|',
    tabs_as_spaces(string, tab.stops=c(2, 3, 8))
) )
}
\seealso{
\link{fansi} for important details on how strings are interpreted,
particularly if you are getting unexpected results.
}
