% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{term_cap_test}
\alias{term_cap_test}
\title{Test Terminal Capabilities}
\usage{
term_cap_test()
}
\value{
character the test vector, invisibly
}
\description{
Outputs ANSI CSI SGR formatted text to screen so that you may visually
inspect what color capabilities your terminal supports.  The supported styles
should be rendered in yellow text with a blue background.
}
\examples{
term_cap_test()
}
